requirements.txt
fastapi
uvicorn[standard]
google-cloud-bigquery
google-cloud-aiplatform
vertexai
pydantic
